Small fixes and refactoring
